feature: TF2 SDK + 64Bit port (#1)
* root: TF2 SDK + 64Bit port

* git: Update .gitignore

* public, vgui: Remove some DI leftovers; Remove VGui2 DLL sources

+++ src/hammer/TextureSystem.cpp
@@ -57,7 +57,7 @@ CTextureSystem g_Textures;
 //-----------------------------------------------------------------------------
 // CMaterialFileChangeWatcher implementation.
 //-----------------------------------------------------------------------------
-void CMaterialFileChangeWatcher::Init( CTextureSystem *pSystem, int context )
+void CMaterialFileChangeWatcher::Init( CTextureSystem *pSystem, intp context )
 {
 	m_pTextureSystem = pSystem;
 	m_Context = context;
@@ -608,11 +608,11 @@ void CTextureSystem::LoadMaterials(CGameConfig *pConfig)
 	m_pActiveContext->Groups.AddToTail(pGroup);
 
 	// Add all the materials to the group.
-	CMaterial::EnumerateMaterials( this, "materials", (int)pGroup, INCLUDE_WORLD_MATERIALS );
+	CMaterial::EnumerateMaterials( this, "materials", (intp)pGroup, INCLUDE_WORLD_MATERIALS );
 	
 	// Watch the materials directory recursively...
 	CMaterialFileChangeWatcher *pWatcher = new CMaterialFileChangeWatcher;
-	pWatcher->Init( this, (int)pGroup );
+	pWatcher->Init( this, (intp)pGroup );
 	m_ChangeWatchers.AddToTail( pWatcher );
 
 	Assert( m_pCubemapTexture == NULL );
@@ -1290,7 +1290,7 @@ bool CTextureSystem::HasTexturesForConfig(CGameConfig *pConfig)
 //-----------------------------------------------------------------------------
 // Used to add all the world materials into the material list
 //-----------------------------------------------------------------------------
-bool CTextureSystem::EnumMaterial( const char *pMaterialName, int nContext )
+bool CTextureSystem::EnumMaterial( const char *pMaterialName, intp nContext )
 {
 	CTextureGroup *pGroup = (CTextureGroup *)nContext;
 	CMaterial *pMaterial = CMaterial::CreateMaterial(pMaterialName, false);
